pyproject.toml
[tool.poetry]
name = "vtrace"
version = "0.1.0"
description = ""
authors = ["Richard Bieringa <richardbieringa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
scapy = "^2.4.5"
click = "^8.0.3"
ipinfo = "^4.2.1"
python-dotenv = "^0.19.2"

[tool.poetry.scripts]
cli = "vtrace.app:main"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-cov = "^3.0.0"
pytest-xdist = "^2.5.0"
black = {version = "^21.12b0", allow-prereleases = true}
flake8 = "^4.0.1"
isort = "^5.10.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
